# swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "NO SQL POC",
      "description": "Profile CRUD",
      "version": "1.0.0"
    },
...
    "servers": [
      {
        "url": "/",
        "description": "local"
      }
    ],
    "paths": {
...
      "/profiles": {
        "post": {
          "tags": [
            "profile"
          ],
          "summary": "profile create",
          "description": "user profile create",
          "operationId": "profile",
          "parameters": [
            {
              "$ref": "#/components/securitySchemes/AccessToken"
            }
          ],
          "requestBody": {
            "description": "profile",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/profileRequest"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Successfully created user profiles"
            },
            "400": {
              "$ref": "#/components/responses/InvalidRequestError"
            },
            "401": {
              "$ref": "#/components/responses/UnauthorizedError"
            },
            "500": {
              "$ref": "#/components/responses/ServerError"
            }
          }
        },
        "get": {
          "tags": [
            "profile"
          ],
          "summary": "Get all profiles",
          "description": "Retrieve a profiles",
          "operationId": "getProfileAll",
          "parameters": [
            {
              "$ref": "#/components/securitySchemes/AccessToken"
            }
          ],
          "responses": {
            "200": {
              "description": "Successfully retrieved profiles by color",
              "type": "object"
            },
            "400": {
              "$ref": "#/components/responses/InvalidRequestError"
            },
            "401": {
              "$ref": "#/components/responses/UnauthorizedError"
            },
            "404": {
              "$ref": "#/components/responses/NotFoundError"
            },
            "500": {
              "$ref": "#/components/responses/ServerError"
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes": {
        "AccessToken": {
          "name": "x-access-token",
          "in": "header",
          "description": "Access Token",
          "required": true,
          "schema":{
            "type": "string"
          }
        }
      },
      "schemas": {
        "profileRequest": {
          "type": "object"
        },
...
        "BasicErrorMessage": {
          "type": "object",
          "required": [
            "error"
          ],
          "properties": {
            "error": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "responses": {
        "UnauthorizedError": {
          "description": "Logged in user is unauthorized to make this request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicErrorMessage"
              }
            }
          }
        },
        "InvalidRequestError": {
          "description": "Missing required fields in request body",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicErrorMessage"
              }
            }
          }
        },
        "NotFoundError": {
          "description": "Resource was not found",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicErrorMessage"
              }
            }
          }
        },
...
        "ServerError": {
          "description": "Unexpected server error",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicErrorMessage"
              }
            }
          }
        }
      }
    }
  }